requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo>=4.3.3
scikit-learn
pyaml
mlflow
fastapi
uvicorn
scipy
apache-airflow
Jinja2
streamlit
requests 
beautifulsoup4 
python-whois 
apache-airflow-providers-mongo

# -e .